Removed personal domain leftovers
There was still a hardcoded personal domain inside the template, which
is now removed and the zone email address is now configurable. By
default it will be zoneadmin.firstdnsserverdomain, but you can set it
either in the template per zone or just set a general address using the
variable sssdns_zoneadmin in your plays.