Fixed logical error for dynamic zones
The variable which is set for a detected dynamic zone (a zone where
a TSIG key for dynamic updates exists as well) is used in a loop and
not reset. So if zone x was dynamic, and the upcoming zone was not,
the variable was still set and could lead to issues updating the
zones. So now the detection registers its own variable and the check
looks if the task was skipped (not a dynamicly managed zone) or not.
So it should work now as expected :)