Add: DNSSEC support
As knot supports automatic DNSSEC support, i wanted to add this to my
role as well. Its not much that needs to be done in general, but it
requires some rework in the role for the active zones tracking.
Instead of just a plain list of active zones, we now have a list with
the DNSSEC options. This might also allow future per-domain options
if needed.
Many thanks to Janine for some help regarding the active zone
tracking :)