New: Container Update and TAGs Support
This update adds support to update the knotd conainer and for that
reason also introduces tags to the role. Tags are not completed
yet as its a bit more complicated than expected. By now the
containerupdate and zones tag exist, while zones somehow struggle
a bit every now and than. Also check mode is a bit hard, but
mostly working currently.