Fix: changed name<>ip mapping for dns server hostnames
The default zone was adding the names of the dns servers to the
dns main zones. So if a servers name was "server-a" it was added
as server-a.dnszonename.tld, but with ansibles default detected
IP. But at least for some cloud servers like Hetzner, the IP was
wrongly detected if you had floating IPs configured as well.
So instead of guessing the right IP, we now have a variable to
configure it (sss_mgmt_addr_v4+sss_mgmt_addr_v6).