Fix: sorting added to template
As i tried to deploy from multiple systems, i noticed the order in
the host groups for example is not strictly the same. For that
reason i added a few sort's to the template to ensure it always
renders the same. Not 100% if i catched all variations that might
cause diffs, but at least for my tests it seems to sort out the
issue ;)