Update: use gathered vars instead of command
While doing some testing, i saw that its not really needed to exec
the date command to get the epoch time. When initially developing,
the facts gathering wasn't always enabled, so i never thought about
it. But this is long gone, so lets get rid of the task and make it
even a bit faster and drop the dependency on "date" ;)